+++ filmeet/src/main/java/com/ureca/filmeet/domain/movie/controller/query/MovieQueryController.java
@@ -10,8 +10,8 @@ import com.ureca.filmeet.domain.movie.dto.response.MoviesResponse;
 import com.ureca.filmeet.domain.movie.dto.response.MoviesSearchByGenreResponse;
 import com.ureca.filmeet.domain.movie.dto.response.RecommendationMoviesResponse;
 import com.ureca.filmeet.domain.movie.dto.response.UpcomingMoviesResponse;
-import com.ureca.filmeet.domain.movie.repository.cache.BoxOfficeRedisQueryService;
 import com.ureca.filmeet.domain.movie.repository.querydsl.SliceWithCount;
+import com.ureca.filmeet.domain.movie.service.query.BoxOfficeRedisQueryService;
 import com.ureca.filmeet.domain.movie.service.query.MovieQueryService;
 import com.ureca.filmeet.domain.movie.service.query.MovieRecommendationQueryService;
 import com.ureca.filmeet.domain.movie.service.query.MovieUpcomingQueryService;
